moved remove_old_log_files to utils
